Improve test coverage of exercises.Fibonacci.
Added a properties check to ensure that for all n > 2, Fibonacci.fib
satisfies the invariant fib(n) := fib(n - 1) + fib(n - 2).

Testing large values of n returned by the generator revealed that
the naive implementation of fib was too inefficient, so I implemented
a linear time version which now runs in a reasonable amount of time.